requirements.txt
requests==2.25.1
discord.py==1.7.2
python-dotenv==0.17.1